Generalize plotted region
This removes assumptions from the `grid-plot` component that the
supplied line data forms a grid within the unit square. Now, the extent
of the region displayed will be computed from the supplied line data.

This commit also removes support for displaying points on the grid. They
may come back in the future, but for now are a superfluous complication.