Refactor point representation
Instead of labeling real points with { x, y } objects and complex points
with { re, im } objects, all points are represented agnostically as [a,
b] arrays. In many cases we were treating complex points as real points
in order to visualize them, so this removes the need for lots of
branching logic.